requirement.txt
flask
pdfplumber
sklearn_crfsuite
scikit-learn<=0.24
nltk
ray
pandas
flask-cors